Extract js related functions from MongoDbPlanner into separate file
